requirements.txt
streamlit>=1.31.0
anthropic>=0.18.0
openai>=1.12.0
python-dotenv>=1.0.0